Crate clockwork_base32

Clockwork Base32 decoding and encoding

This module contains high level functions and low level functions for decoding and encoding bytes with Clockwork Base32.

Examples

You can encode bytes to a String with encode_to_string:

use clockwork_base32 as base32;
let encoded = base32::encode_to_string(b"Hello, world!");
assert_eq!(&encoded, "91JPRV3F5GG7EVVJDHJ22");

You can decode bytes to a String with decode_to_string:

use clockwork_base32 as base32;
let decoded = base32::decode_to_string(b"91JPRV3F5GG7EVVJDHJ22")?;
assert_eq!(&decoded, "Hello, world!");

High level functions

These functions decode/encode bytes and return a new String or Vec<u8> as the result.

decode_to_string

encode_to_string

decode_to_vec

encode_to_vec

Low level functions

These functions take a String or Vec<u8> argument for the destination and append the decoded/encoded result to it.

append_decoded_to_string

append_decoded_to_vec

append_encoded_to_string

append_encoded_to_vec

These functions can be used to calculate the capacity for the decode/encode result beforehand.

capacity_hint_for_decode

capacity_hint_for_encode

Functions

append_decoded_to_string
Decodes bytes and append the result to dest.

append_decoded_to_vec
Decodes bytes and append the result to dest.

append_encoded_to_string
Encodes bytes and append the result to dest.

append_encoded_to_vec
Encodes bytes and append the result to dest.

capacity_hint_for_decode
Returns a hint for the capacity needed for the decoded result.

capacity_hint_for_encode
Returns a hint for the capacity needed for the encoded result.

decode_to_string
Decodes bytes and returns the result as a new String.

decode_to_vec
Decodes bytes and returns the result as a new Vec<u8>.

encode_to_string
Encodes bytes and returns the result as a new String.

encode_to_vec
Encodes bytes and returns the result as a new Vec<u8>.